Added unit testing of admin api,created navbar,created admin feedback

+++ FruitAdminApi/Repository/AdminFruitRepository.cs
@@ -1,11 +1,13 @@
 ﻿using FruitUserApi.Data;
 using FruitUserApi.Models;
 using System.Web;
+using FruitAdminApi.Custom_Exceptions;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using System.IO;
+using System;
 
 namespace FruitAdminApi.Repository
 {
@@ -23,15 +25,32 @@ namespace FruitAdminApi.Repository
             return db.Fruits.ToList();
         }
         //Get Fruit Details By id
-        public Fruit GetFruitById(int id)
+        public async Task<Fruit> GetFruitById(int? id)
         {
-            Fruit fruit = db.Fruits.Find(id);
+            if(id==null)
+            {
+                throw new NullValueException("FruitId cant be empty");
+            }
+            Fruit fruit = await db.Fruits.FindAsync(id);
+            if(fruit==null)
+            {
+                throw new NotFoundException("Fruit not found");
+            }
             return fruit;
         }
         //edit fruit by id
-        public async Task<Fruit> EditFruitByIdAsync(int id,Fruit model)
+        public async Task<Fruit> EditFruitByIdAsync(int? id,Fruit model)
         {
-            Fruit fruit = db.Fruits.Find(id);
+            if (id == null)
+            {
+                throw new NullValueException("FruitId cant be empty");
+            }
+            
+            Fruit fruit = await db.Fruits.FindAsync(id);
+            if (fruit == null)
+            {
+                throw new NotFoundException("Fruit not found");
+            }
             fruit.FruitName = model.FruitName;
             fruit.FruitImg = model.FruitImg;
             fruit.FruitPrice = model.FruitPrice;
@@ -41,8 +60,12 @@ namespace FruitAdminApi.Repository
         }
 
         // delete Fruit By id
-        public async Task<bool> DeleteFruitByIdAsync(int id)
+        public async Task<bool> DeleteFruitByIdAsync(int? id)
         {
+            if (id == null)
+            {
+                throw new NullValueException("FruitId cant be empty");
+            }
             Fruit fruit = db.Fruits.Find(id);
             db.Fruits.Remove(fruit);
             await db.SaveChangesAsync();
@@ -53,6 +76,14 @@ namespace FruitAdminApi.Repository
         //Add Fruit
         public async Task<Fruit> AddFruitAsync(Fruit model)
         {
+            List<Fruit> FruitList = GetAllFruits();
+            foreach (var fruit in FruitList)
+            {
+                if ((model.FruitName).Equals((fruit.FruitName),StringComparison.OrdinalIgnoreCase)) 
+                {
+                    throw new AlreadyExistsException(" This Fruit name already Exists!");
+                }
+            }
             db.Fruits.Add(model);
             await db.SaveChangesAsync();
             return model;
